Add in a custom pluggable Header
This changes how the Header offset works. This will allow the user to
reserve the first page for either the ringbuffer cursor (cool!), or a
custom header, which returns a pointer to the (opaque!) diskring.Cursor.

We don't want users to play with our state, so we don't export the
fields, so they have to treat it like an opaque struct of an unknown
size. My guess is most users will want to stuff this at the end of their
struct in order to keep ABI stable for themselves. That seems fine to
mee, we have 4k to use.